requirements.txt
fastapi[all]
uvicorn[standard]
granian
sqlalchemy
psycopg2-binary
pydantic-settings
GeoAlchemy2
svix
python-jose[cryptography]
sqladmin
requests

# for tests
pytest
pytest-cov
httpx
pytest-html
pytest-json-report
ruff
mypy
bandit
safety
locust
